design(Layout) Redesigned Layout
This CL updates the overall look and feel of my homepage. This redesign
simplifies the layout, removing a lot of the text and instead just displaying
my name and logos of companies I've worked at.

As fun as it is to talk about what I worked on... I have linkedin for that.

For now its a very empty webpage, which I kind of like. I might put back the
Game of Life rendering again... or maybe add a game in the empty space... it
all depends on how I'm feeling. :)